pyproject.toml
[tool.poetry]
name = "liza_cli"
version = "0.1.0"
description = "A BitBucket PR tracking CLI"
authors = ["Chris Muthig <camuthig@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
typer-cli = "^0.0.9"
httpx = "^0.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
